Updated hmc_type from uint8_t to uint32_t to support new memH format. - hmc_trans is not used anymore - should use memH new functions to set hmc_type (memFlags in memH) -- in next commit

+++ macsimComponent.h
@@ -100,7 +100,7 @@ private:
     // callback functions for sending requests; these will be called by macsim
     void sendInstructionCacheRequest(int,uint64_t,uint64_t,int);
 #ifdef USE_VAULTSIM_HMC
-    void sendDataCacheRequest(int,uint64_t,uint64_t,int,int,uint8_t,uint64_t);
+    void sendDataCacheRequest(int,uint64_t,uint64_t,int,int,uint32_t,uint64_t);
 #else
     void sendDataCacheRequest(int,uint64_t,uint64_t,int,int);
 #endif
